Refactored PidAwareTextTestResult and fixed pr/prhead prefixes
